Add Rules page to website frontend
