+++ CelesteNet.Server.FrontendModule/Content/frontend/index.html
@@ -166,11 +166,12 @@
                             CelesteNet is an <b>online multiplayer mod for <a href="http://www.celestegame.com/">Celeste</a></b>, using the <a href="https://everestapi.github.io/">Everest mod loader</a>, made with a lot of love and sweat.<br>
                             It lets you join levels together with your friends or complete strangers, even jumping on others' heads and throwing them across pits if you want to!
                             There even are some <a href="https://gamebanana.com/mods/150725">multiplayer-exclusive maps</a> and <a href="https://gamebanana.com/mods/335146">custom game modes</a> for you to experience already, and more in development by the community.<br>
+                            <h3>⚠️ Please read and follow the <strong><a href="/rules">CelesteNet Rules</a></strong> when connecting to CelesteNet.</h3>
                         </p>
                     </div>
                     <div>
                         <h2>How do I play on CelesteNet?</h2>
-                        </p>
+                        <p>
                             First, make sure that you've installed <a href="https://everestapi.github.io/">Everest (loader) via Olympus (manager)</a>.<br>
                             Click on the "Download CelesteNet" button above, then click the "1-CLICK INSTALL" button on the latest version.
                             <h3>Once you're in-game:</h3>
@@ -185,7 +186,7 @@
                     <div>
                         <h2>Boring stuff:</h2>
                         <p>
-                            The CelesteNet server is hosted by <a href="https://0x0ade.github.io/">0x0ade</a>. The <a href="https://github.com/0x0ade/CelesteNet"></a>source code is available on GitHub</a> and is being maintained by multiple people.<br>
+                            The CelesteNet server is hosted by <a href="https://0x0ade.github.io/">0x0ade</a>. The <a href="https://github.com/0x0ade/CelesteNet">source code is available on GitHub</a> and is being maintained by multiple people.<br>
                             Hosting your own server is possible, but not advised and is very "do it yourself." Instead, it's much easier to create a private channel (/help join).<br>
                             <br>
                             Connecting to any server will share some basic information (IP address, nickname, messages, ...), which then gets logged temporarily to prevent abuse and to ban spammers and other misbehaving players.<br>
@@ -232,6 +233,7 @@
             window.netdataNoBootstrap = true;
         </script>
         <script src="/main/index.js" type="module" async></script>
+        <script src="/main/clipswrap.js" type="module" async></script>
         <script src="https://netdata.0x0a.de/dashboard.js" async defer></script>
         <script async defer>
             function netdataHook() {
