+++ CelesteNet.Server.FrontendModule/Content/frontend/rules/index.html
@@ -1,0 +1,111 @@
+﻿<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
+    <link rel="icon" type="image/png" href="/favicon.png">
+    <link rel="shortcut icon" type="image/png" href="/favicon.png">
+
+    <title>CelesteNet Rules</title>
+
+    <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
+
+    <meta name="mobile-web-app-capable" content="yes">
+    <meta name="theme-color" content="#3b2d4a">
+
+    <meta property="og:title" content="CelesteNet Rules" />
+    <meta name="description" content="Community Rules for the online multiplayer mod CelesteNet.">
+    <meta property="og:description" content="Community Rules for the online multiplayer mod CelesteNet.">
+    <meta name="author" content="0x0ade">
+    <meta property="og:locale" content="en_US">
+
+    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Poppins:400,600&display=swap">
+
+    <link rel="stylesheet" href="/main/style.css">
+    <link rel="stylesheet" href="/overlay/style.css">
+</head>
+<body>
+    <div id="background">
+        <div id="cogwheel"></div>
+    </div>
+
+    <div id="body">
+        <div id="front" class="rules-front">
+            <div id="front-clips-wrap">
+                <video id="front-clips-video" src="https://papyrus.0x0a.de/celestenet/clips_opt.mp4" width="484" height="272" autoplay="autoplay" loop="true" muted="true"></video>
+            </div>
+            <div id="front-dim"></div>
+            <div id="front-title">CelesteNet Rules</div>
+            <div id="front-text">Community Rules for the online multiplayer mod <a href="/">CelesteNet</a>.</div>
+        </div>
+
+        <div id="main" class="rules-main">
+
+            <div id="content">
+                <div>
+                    <sub class="back-link"><a href="/">&lt; Back to main page</a></sub>
+                    <h2>CelesteNet rules? Heck yeah it does!</h2>
+                    <p>
+                        Jokes aside, let's be real &mdash; any community needs some rules and guidelines, to make sure everyone is having fun and feels welcome to participate.<br />
+                        That is why we've adopted/adapted the rules of the <a href="https://discord.gg/celeste">official Celeste Discord server</a> to cover CelesteNet as well.<br />
+                        <br />
+                        These rules apply to <strong>global chat</strong> and the default channel of "<strong>main</strong>". For other channels, especially private ones and whispers, we won't read every last detail of
+                        your private conversations among friends. But please try to also stick to the rules there.<br />
+                        <br />
+                        Failure to follow the rules <strong>will</strong> result in warnings, kicks and ultimately bans where necessary, as deemed necessary by the Team.<br />
+                        <br />
+                        We, the CelesteNet Team, try to always have an eye on what's going on in CelesteNet and its global chat, but we can't guarantee to always catch everything as it happens.<br />
+                        So if you see someone breaking these rules, something makes you uncomfortable or someone is trying to ruin your CelesteNet experience, please report it to us in the <strong>#celestenet</strong> channel on
+                        the <a href="https://discord.gg/celeste">Celeste Discord</a>.<br />
+                        You can also contact one of the members with the <strong>@CelesteNet Team</strong> role in DMs if you want.
+                    </p>
+                </div>
+                <div id="rules">
+                    <h1>CelesteNet Rules</h1>
+                    <ol>
+                        <li>
+                            Don't harass or discriminate against others, including using slurs or words in any way that puts down a specific group of people, even when not used to refer to those groups. Changing or censoring a banned word in an attempt to bypass this rule is prohibited.
+                        </li>
+                        <li>
+                            Discussion about NSFW, disturbing, or malicious content is not allowed.
+                        </li>
+                        <li>
+                            Don't cause or contribute to drama. This includes discussion about politics, religion, gossip about other people, and any other controversial topics, as they often cause unnecessary arguments.
+                        </li>
+                        <li>
+                            No spamming. as posting chains of emotes, excessively long text, and gibberish.
+                        </li>
+                        <li>
+                            Do not vent or make the general topic of discussion about your life problems or actions of self harm.
+                            If you are having thoughts about self harm it is important to talk to someone trained to help you, so be sure to look into local resources for people who can help you help yourself.
+                        </li>
+                        <li>
+                            Be respectful and understanding of others, and avoid making people uncomfortable or otherwise ruining their fun. We all get upset at times, but negativity should stay at a reasonable level, as it drains everyone else.
+                        </li>
+                        <li>
+                            If someone is posting a pb/accomplishment in gaming or life, either congratulate them or keep silent, unless constructive criticism or feedback is requested. No one is claiming that their achievements are unbeatable, anything could be faster, so this doesn't need to be stated.
+                        </li>
+                        <li>
+                            Community members should strive to make others feel comfortable with their achievements without pushing them to do more beyond that. Peer pressure, valuing others' personal achievements/feelings as less than yours, or one-upping others with mentions aren't allowed.
+                        </li>
+                        <li>
+                            User profiles, including but not limited to Username, Nickname, and Profile Picture are subject to moderation in accordance to the rules, in particular rules 1 and 2.
+                        </li>
+                        <li>
+                            Please only use English for conversations in global chat and the "main" channel. As extended discussion in other languages is difficult to moderate.
+                        </li>
+                    </ol>
+                </div>
+
+                <div id="baked">
+                    Baked with <img class="berry" src="/main/berry.gif" alt="a strawberry"> by <a href="https://github.com/0x0ade/CelesteNet">the CelesteNet devs</a>
+                </div>
+            </div>
+        </div>
+    </div>
+
+    <div id="dummy"></div>
+
+    <script src="/main/clipswrap.js" type="module" async></script>
+</body>
+</html>
